push_notifications: Adjust APNs tokens to be case-insensitive in the database.
APNs apparently treats its tokens case-insensitively; FCM does not.
Adjust the `unique_together` to instead be separate partial
constraints, keyed on the `kind` of the PushDeviceToken.

+++ zilencer/views.py
@@ -12,8 +12,9 @@ from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.core.validators import URLValidator, validate_email
 from django.db import IntegrityError, transaction
-from django.db.models import Model
+from django.db.models import Model, QuerySet
 from django.db.models.constants import OnConflict
+from django.db.models.functions import Lower
 from django.http import HttpRequest, HttpResponse
 from django.utils.crypto import constant_time_compare, get_random_string
 from django.utils.timezone import now as timezone_now
@@ -424,6 +425,26 @@ def check_transfer_challenge_response_secret_not_prepared(response: requests.Res
     return secret_not_prepared
 
 
+def get_remote_push_device_token(
+    *,
+    server: RemoteZulipServer,
+    token: str,
+    kind: int,
+) -> QuerySet[RemotePushDeviceToken]:
+    if kind == RemotePushDeviceToken.APNS:
+        return RemotePushDeviceToken.objects.alias(lower_token=Lower("token")).filter(
+            server=server,
+            lower_token=token.lower(),
+            kind=kind,
+        )
+    else:
+        return RemotePushDeviceToken.objects.filter(
+            server=server,
+            token=token,
+            kind=kind,
+        )
+
+
 @typed_endpoint
 def register_remote_push_device(
     request: HttpRequest,
@@ -446,9 +467,11 @@ def register_remote_push_device(
         kwargs: dict[str, object] = {"user_uuid": user_uuid, "user_id": None}
         # Delete pre-existing user_id registration for this user+device to avoid
         # duplication. Further down, uuid registration will be created.
-        RemotePushDeviceToken.objects.filter(
-            server=server, token=token, kind=token_kind, user_id=user_id
-        ).delete()
+        get_remote_push_device_token(
+            server=server,
+            token=token,
+            kind=token_kind,
+        ).filter(user_id=user_id).delete()
     else:
         # One of these is None, so these kwargs will lead to a proper registration
         # of either user_id or user_uuid type
@@ -622,9 +645,11 @@ def unregister_remote_push_device(
 
     update_remote_realm_last_request_datetime_helper(request, server, realm_uuid, user_uuid)
 
-    (num_deleted, ignored) = RemotePushDeviceToken.objects.filter(
-        user_identity.filter_q(), token=token, kind=token_kind, server=server
-    ).delete()
+    (num_deleted, ignored) = (
+        get_remote_push_device_token(token=token, kind=token_kind, server=server)
+        .filter(user_identity.filter_q())
+        .delete()
+    )
     if num_deleted == 0:
         raise JsonableError(err_("Token does not exist"))
 
@@ -683,7 +708,10 @@ def delete_duplicate_registrations(
     assert len({registration.kind for registration in registrations}) == 1
     kind = registrations[0].kind
 
-    tokens_counter = Counter(device.token for device in registrations)
+    if kind == RemotePushDeviceToken.APNS:
+        tokens_counter = Counter(device.token.lower() for device in registrations)
+    else:
+        tokens_counter = Counter(device.token for device in registrations)
 
     tokens_to_deduplicate = []
     for key in tokens_counter:
@@ -757,11 +785,12 @@ def remote_server_send_test_notification(
 
     update_remote_realm_last_request_datetime_helper(request, server, realm_uuid, user_uuid)
 
-    try:
-        device = RemotePushDeviceToken.objects.get(
-            user_identity.filter_q(), token=token, kind=token_kind, server=server
-        )
-    except RemotePushDeviceToken.DoesNotExist:
+    device = (
+        get_remote_push_device_token(token=token, kind=token_kind, server=server)
+        .filter(user_identity.filter_q())
+        .first()
+    )
+    if device is None:
         raise InvalidRemotePushDeviceTokenError
 
     send_test_push_notification_directly_to_devices(
@@ -1046,16 +1075,38 @@ def get_deleted_devices(
         kind=RemotePushDeviceToken.FCM,
         server=server,
     ).values_list("token", flat=True)
-    apple_devices_we_have = RemotePushDeviceToken.objects.filter(
-        user_identity.filter_q(),
-        token__in=apple_devices,
-        kind=RemotePushDeviceToken.APNS,
-        server=server,
-    ).values_list("token", flat=True)
+
+    # APNS tokens are case-insensitive -- but the remote server may
+    # not know that yet.  As such, we perform our local lookups
+    # case-insensitively, returning the exact case the remote server
+    # used, and also return all-but-one of any case duplicates that
+    # the remote server passed us.
+    canonical_case = {}
+    apns_token_to_remove = set()
+    for token in apple_devices:
+        if token.lower() not in canonical_case:
+            canonical_case[token.lower()] = token
+        elif canonical_case[token.lower()] == token:
+            # Be careful to skip if identical-case tokens somehow show up more than once
+            pass
+        else:
+            apns_token_to_remove.add(token)
+    apple_devices_we_have = (
+        RemotePushDeviceToken.objects.annotate(lower_token=Lower("token"))
+        .filter(
+            user_identity.filter_q(),
+            lower_token__in=canonical_case.keys(),
+            kind=RemotePushDeviceToken.APNS,
+            server=server,
+        )
+        .values_list("lower_token", flat=True)
+    )
+    for token_to_remove in set(canonical_case.keys()) - set(apple_devices_we_have):
+        apns_token_to_remove.add(canonical_case[token_to_remove])
 
     return DevicesToCleanUpDict(
-        android_devices=list(set(android_devices) - set(android_devices_we_have)),
-        apple_devices=list(set(apple_devices) - set(apple_devices_we_have)),
+        android_devices=sorted(set(android_devices) - set(android_devices_we_have)),
+        apple_devices=sorted(apns_token_to_remove),
     )
